requirements.txt
boto3==1.35.76
mypy-boto3==1.35.70
boto3-stubs[s3,glue,athena]
pandas==2.2.3
s3fs==0.4.2
polars==1.16.0
pyparsing
moto
pytest